Completes python 3 tests
Issues:
Fixes #287

Problem:
codebase needed to be python3 compat

Analysis:
Completes python 3 compat

Tests:
