+++ .gitignore
@@ -30,6 +30,7 @@ develop-eggs
 lib
 lib64
 __pycache__
+**/config.JSON
 
 # Installer logs
 pip-log.txt
